Added addProduct test Modified App.main to better fit junit tests
